// src/app/api-client/model/signup.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Signup { 
    user_first_name: string;
    user_last_name: string;
    user_email: string;
    user_password: string;
    user_age: number | null;
    user_city: string | null;
    user_country: string | null;
    user_gender: string | null;
    user_state: string | null;
}